// src/app/models/class.model.ts
// src/app/models/class.model.ts

export interface Class {
    id: string;         // Identificador único para cada clase
    title: string;      // Título de la clase
    description: string; // Descripción de la clase
    date: string;       // Fecha de la clase, por ejemplo "2024-10-18"
  }
  